backend/requirements.txt
# ============================================================
# 浅愈(GentleMend) Backend — Python 依赖
# ============================================================

# Web 框架
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.1.0,<3.0

# 数据库
sqlalchemy[asyncio]>=2.0.25,<3.0
asyncpg>=0.29.0,<1.0
aiosqlite>=0.20.0,<1.0
alembic>=1.13.0,<2.0

# Redis
redis[hiredis]>=5.0.0,<6.0

# AI API (OpenAI兼容接口，支持DeepSeek/Claude/OpenAI)
openai>=1.30.0,<2.0
httpx>=0.27.0,<1.0

# 环境变量
python-dotenv>=1.0.0,<2.0

# 测试
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
pytest-cov>=4.1.0,<6.0

# 代码质量
ruff>=0.3.0,<1.0
mypy>=1.8.0,<2.0
